Added callback field to GirVariable
