+++ src/types/index.ts
@@ -14,7 +14,6 @@ export * from './gir-construct'
 export * from './gir-doc'
 export * from './gir-enumeration'
 export * from './gir-enumeration-member'
-export * from './gir-function'
 export * from './gir-implements'
 export * from './gir-include'
 export * from './gir-module-resolved-by'
